feat: move exceptions back to common package (#295)
### Summary of Changes

Move exceptions back to common top-level package. This undoes the change
from #177 and prepares for #262 and other errors that don't belong to a
specific package of Safe-DS. It also makes it easier for developer to
find a fitting error.

---------
