+++ src/safeds/data/tabular/containers/_column.py
@@ -12,13 +12,13 @@ import seaborn as sns
 
 from safeds.data.image.containers import Image
 from safeds.data.image.typing import ImageFormat
-from safeds.data.tabular.exceptions import (
+from safeds.data.tabular.typing import ColumnType
+from safeds.exceptions import (
     ColumnLengthMismatchError,
     ColumnSizeError,
     IndexOutOfBoundsError,
     NonNumericColumnError,
 )
-from safeds.data.tabular.typing import ColumnType
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Iterator
