+++ src/safeds/data/tabular/containers/_table.py
@@ -15,7 +15,8 @@ from scipy import stats
 
 from safeds.data.image.containers import Image
 from safeds.data.image.typing import ImageFormat
-from safeds.data.tabular.exceptions import (
+from safeds.data.tabular.typing import ColumnType, Schema
+from safeds.exceptions import (
     ColumnLengthMismatchError,
     ColumnSizeError,
     DuplicateColumnNameError,
@@ -24,7 +25,6 @@ from safeds.data.tabular.exceptions import (
     SchemaMismatchError,
     UnknownColumnNameError,
 )
-from safeds.data.tabular.typing import ColumnType, Schema
 
 from ._column import Column
 from ._row import Row
