--- src/safeds/data/tabular/exceptions/__init__.py
@@ -1,23 +1,0 @@
-"""Custom exceptions that can be raised when working with tabular data."""
-
-from ._exceptions import (
-    ColumnLengthMismatchError,
-    ColumnSizeError,
-    DuplicateColumnNameError,
-    IndexOutOfBoundsError,
-    NonNumericColumnError,
-    SchemaMismatchError,
-    TransformerNotFittedError,
-    UnknownColumnNameError,
-)
-
-__all__ = [
-    "ColumnLengthMismatchError",
-    "ColumnSizeError",
-    "DuplicateColumnNameError",
-    "IndexOutOfBoundsError",
-    "NonNumericColumnError",
-    "SchemaMismatchError",
-    "TransformerNotFittedError",
-    "UnknownColumnNameError",
-]
