+++ src/safeds/data/tabular/transformation/_label_encoder.py
@@ -3,10 +3,10 @@ from __future__ import annotations
 from sklearn.preprocessing import OrdinalEncoder as sk_OrdinalEncoder
 
 from safeds.data.tabular.containers import Table
-from safeds.data.tabular.exceptions import TransformerNotFittedError, UnknownColumnNameError
 from safeds.data.tabular.transformation._table_transformer import (
     InvertibleTableTransformer,
 )
+from safeds.exceptions import TransformerNotFittedError, UnknownColumnNameError
 
 
 # noinspection PyProtectedMember
