+++ src/safeds/exceptions/__init__.py
@@ -1,0 +1,39 @@
+"""Custom exceptions that can be raised by Safe-DS."""
+
+from safeds.exceptions._data import (
+    ColumnLengthMismatchError,
+    ColumnSizeError,
+    DuplicateColumnNameError,
+    IndexOutOfBoundsError,
+    NonNumericColumnError,
+    SchemaMismatchError,
+    TransformerNotFittedError,
+    UnknownColumnNameError,
+)
+from safeds.exceptions._ml import (
+    DatasetContainsTargetError,
+    DatasetMissesFeaturesError,
+    LearningError,
+    ModelNotFittedError,
+    PredictionError,
+    UntaggedTableError,
+)
+
+__all__ = [
+    # Data exceptions
+    "ColumnLengthMismatchError",
+    "ColumnSizeError",
+    "DuplicateColumnNameError",
+    "IndexOutOfBoundsError",
+    "NonNumericColumnError",
+    "SchemaMismatchError",
+    "TransformerNotFittedError",
+    "UnknownColumnNameError",
+    # ML exceptions
+    "DatasetContainsTargetError",
+    "DatasetMissesFeaturesError",
+    "LearningError",
+    "ModelNotFittedError",
+    "PredictionError",
+    "UntaggedTableError",
+]
