+++ src/safeds/exceptions/_data.py
