--- src/safeds/ml/exceptions/__init__.py
@@ -1,19 +1,0 @@
-"""Custom exceptions that can be raised when working with ML models."""
-
-from ._exceptions import (
-    DatasetContainsTargetError,
-    DatasetMissesFeaturesError,
-    LearningError,
-    ModelNotFittedError,
-    PredictionError,
-    UntaggedTableError,
-)
-
-__all__ = [
-    "DatasetContainsTargetError",
-    "DatasetMissesFeaturesError",
-    "LearningError",
-    "ModelNotFittedError",
-    "PredictionError",
-    "UntaggedTableError",
-]
