+++ tests/safeds/data/tabular/containers/_column/test_correlation_with.py
@@ -1,6 +1,6 @@
 import pytest
 from safeds.data.tabular.containers import Column
-from safeds.data.tabular.exceptions import ColumnLengthMismatchError, NonNumericColumnError
+from safeds.exceptions import ColumnLengthMismatchError, NonNumericColumnError
 
 
 @pytest.mark.parametrize(
