+++ tests/safeds/data/tabular/containers/_column/test_plot_boxplot.py
@@ -1,7 +1,7 @@
 import pytest
 from safeds.data.image.containers import Image
 from safeds.data.tabular.containers import Table
-from safeds.data.tabular.exceptions import NonNumericColumnError
+from safeds.exceptions import NonNumericColumnError
 
 from tests.helpers import resolve_resource_path
