+++ tests/safeds/data/tabular/containers/_column/test_stability.py
@@ -2,7 +2,7 @@ from typing import Any
 
 import pytest
 from safeds.data.tabular.containers import Column
-from safeds.data.tabular.exceptions import ColumnSizeError
+from safeds.exceptions import ColumnSizeError
 
 
 @pytest.mark.parametrize(
