+++ tests/safeds/data/tabular/containers/_table/test_add_column.py
@@ -1,6 +1,6 @@
 import pytest
 from safeds.data.tabular.containers import Column, Table
-from safeds.data.tabular.exceptions import ColumnSizeError, DuplicateColumnNameError
+from safeds.exceptions import ColumnSizeError, DuplicateColumnNameError
 
 
 @pytest.mark.parametrize(
