+++ tests/safeds/data/tabular/containers/_table/test_from_dict.py
@@ -2,7 +2,7 @@ from typing import Any
 
 import pytest
 from safeds.data.tabular.containers import Table
-from safeds.data.tabular.exceptions import ColumnLengthMismatchError
+from safeds.exceptions import ColumnLengthMismatchError
 
 
 @pytest.mark.parametrize(
