+++ tests/safeds/data/tabular/containers/_table/test_init.py
@@ -1,7 +1,7 @@
 import pytest
 from safeds.data.tabular.containers import Table
-from safeds.data.tabular.exceptions import ColumnLengthMismatchError
 from safeds.data.tabular.typing import Integer, Schema
+from safeds.exceptions import ColumnLengthMismatchError
 
 
 @pytest.mark.parametrize(
