+++ tests/safeds/data/tabular/containers/_table/test_inverse_transform_table.py
@@ -1,7 +1,7 @@
 import pytest
 from safeds.data.tabular.containers import Table
-from safeds.data.tabular.exceptions import TransformerNotFittedError
 from safeds.data.tabular.transformation import OneHotEncoder
+from safeds.exceptions import TransformerNotFittedError
 
 
 @pytest.mark.parametrize(
