+++ tests/safeds/data/tabular/containers/_table/test_transform_table.py
@@ -1,7 +1,7 @@
 import pytest
 from safeds.data.tabular.containers import Table
-from safeds.data.tabular.exceptions import TransformerNotFittedError, UnknownColumnNameError
 from safeds.data.tabular.transformation import OneHotEncoder
+from safeds.exceptions import TransformerNotFittedError, UnknownColumnNameError
 
 
 @pytest.mark.parametrize(
