+++ tests/safeds/data/tabular/transformation/test_imputer.py
@@ -1,8 +1,8 @@
 import pytest
 from safeds.data.tabular.containers import Table
-from safeds.data.tabular.exceptions import TransformerNotFittedError, UnknownColumnNameError
 from safeds.data.tabular.transformation import Imputer
 from safeds.data.tabular.typing import ImputerStrategy
+from safeds.exceptions import TransformerNotFittedError, UnknownColumnNameError
 
 
 class TestStrategy:
