+++ tests/safeds/data/tabular/typing/test_schema.py
@@ -4,8 +4,8 @@ from typing import TYPE_CHECKING
 
 import pandas as pd
 import pytest
-from safeds.data.tabular.exceptions import UnknownColumnNameError
 from safeds.data.tabular.typing import Boolean, ColumnType, Integer, RealNumber, Schema, String
+from safeds.exceptions import UnknownColumnNameError
 
 if TYPE_CHECKING:
     from typing import Any
