+++ tests/safeds/ml/classical/regression/test_regressor.py
@@ -5,7 +5,15 @@ from typing import TYPE_CHECKING
 import pandas as pd
 import pytest
 from safeds.data.tabular.containers import Column, Table, TaggedTable
-from safeds.data.tabular.exceptions import ColumnLengthMismatchError
+from safeds.exceptions import (
+    ColumnLengthMismatchError,
+    DatasetContainsTargetError,
+    DatasetMissesFeaturesError,
+    LearningError,
+    ModelNotFittedError,
+    PredictionError,
+    UntaggedTableError,
+)
 from safeds.ml.classical.regression import (
     AdaBoost,
     DecisionTree,
@@ -22,14 +30,6 @@ from safeds.ml.classical.regression import (
 
 # noinspection PyProtectedMember
 from safeds.ml.classical.regression._regressor import _check_metrics_preconditions
-from safeds.ml.exceptions import (
-    DatasetContainsTargetError,
-    DatasetMissesFeaturesError,
-    LearningError,
-    ModelNotFittedError,
-    PredictionError,
-    UntaggedTableError,
-)
 
 if TYPE_CHECKING:
     from _pytest.fixtures import FixtureRequest
